+++ packages/plugin-packages/mashroom-portal-admin-app/src/js/types.ts
@@ -30,6 +30,7 @@ export type State = {
     readonly selectedPage: SelectedPage | undefined | null;
     readonly selectedSite: SelectedSite | undefined | null;
     readonly modals: ModalState;
+    readonly appWrapperDataAttributes: Record<string,string>;
     readonly tabDialogs: TabDialogState;
 }